[ast][compiler][optimization] Eliminate all MIR related code
## Summary

HIR optimzation is now productionized. It's the time to nuke the MIR code.

## Test Plan

`yarn test`
