--- samlang-core-compiler/__tests__/mir-basic-block-reorder.test.ts
@@ -1,131 +1,0 @@
-/* eslint-disable camelcase */
-
-import createMidIRBasicBlocks from '../mir-basic-block';
-import reorderMidIRBasicBlocksToMaximizeLongestNoJumpPath from '../mir-basic-block-reorder';
-import MidIRResourceAllocator from '../mir-resource-allocator';
-
-import {
-  MidIRStatement_DANGEROUSLY_NON_CANONICAL,
-  midIRStatementToString,
-  MIR_ZERO,
-  MIR_TEMP,
-  MIR_JUMP,
-  MIR_MOVE_TEMP,
-  MIR_LABEL,
-  MIR_CJUMP_NON_FALLTHROUGH_NON_CANONICAL,
-  MIR_RETURN,
-} from 'samlang-core-ast/mir-nodes';
-
-const reorderAndDumpToString = (
-  statements: readonly MidIRStatement_DANGEROUSLY_NON_CANONICAL[]
-): string =>
-  reorderMidIRBasicBlocksToMaximizeLongestNoJumpPath(
-    createMidIRBasicBlocks(new MidIRResourceAllocator(), '', statements)
-  )
-    .map((it) => it.allStatements)
-    .flat()
-    .map(midIRStatementToString)
-    .join('\n');
-
-it('reorderMidIRBasicBlocksToMaximizeLongestNoJumpPath empty test', () => {
-  expect(reorderAndDumpToString([])).toBe('');
-});
-
-it('reorderMidIRBasicBlocksToMaximizeLongestNoJumpPath single statement self-consistency test', () => {
-  expect(reorderAndDumpToString([MIR_RETURN(MIR_ZERO)])).toBe(
-    'l__0_BASIC_BLOCK_1ST_STMT:\nreturn 0;'
-  );
-});
-
-it('reorderMidIRBasicBlocksToMaximizeLongestNoJumpPath simple stream test.', () => {
-  expect(
-    reorderAndDumpToString([
-      MIR_LABEL('a'),
-      MIR_LABEL('b'),
-      MIR_LABEL('c'),
-      MIR_LABEL('d'),
-      MIR_RETURN(MIR_ZERO),
-    ])
-  ).toBe(`a:
-goto b;
-b:
-goto c;
-c:
-goto d;
-d:
-return 0;`);
-});
-
-it('reorderMidIRBasicBlocksToMaximizeLongestNoJumpPath heavy true branch test.', () => {
-  expect(
-    reorderAndDumpToString([
-      MIR_CJUMP_NON_FALLTHROUGH_NON_CANONICAL(MIR_TEMP(''), 'a', 'end'),
-      MIR_LABEL('end'),
-      MIR_JUMP('program_end'),
-      MIR_LABEL('a'),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_JUMP('end'),
-      MIR_LABEL('program_end'),
-      MIR_RETURN(MIR_ZERO),
-    ])
-  ).toBe(`l__0_BASIC_BLOCK_1ST_STMT:
-if () goto a; else goto end;
-a:
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-goto end;
-end:
-goto program_end;
-program_end:
-return 0;`);
-});
-
-it('reorderMidIRBasicBlocksToMaximizeLongestNoJumpPath heavy false branch test.', () => {
-  expect(
-    reorderAndDumpToString([
-      MIR_CJUMP_NON_FALLTHROUGH_NON_CANONICAL(MIR_TEMP(''), 'end', 'a'),
-      MIR_LABEL('end'),
-      MIR_JUMP('program_end'),
-      MIR_LABEL('a'),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_MOVE_TEMP('a', MIR_TEMP('a')),
-      MIR_JUMP('end'),
-      MIR_LABEL('program_end'),
-      MIR_RETURN(MIR_ZERO),
-    ])
-  ).toBe(`l__0_BASIC_BLOCK_1ST_STMT:
-if () goto end; else goto a;
-a:
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-a = a;
-goto end;
-end:
-goto program_end;
-program_end:
-return 0;`);
-});
