+++ samlang-core-compiler/index.ts
@@ -1,9 +1,4 @@
 import compileSamlangSourcesToHighIRSources from './hir-toplevel-lowering';
 import lowerHighIRModuleToLLVMModule from './llvm-lowering-translator';
-import compileHighIrModuleToMidIRCompilationUnit from './mir-toplevel-lowering';
 
-export {
-  compileSamlangSourcesToHighIRSources,
-  compileHighIrModuleToMidIRCompilationUnit,
-  lowerHighIRModuleToLLVMModule,
-};
+export { compileSamlangSourcesToHighIRSources, lowerHighIRModuleToLLVMModule };
