--- samlang-core-interpreter/__tests__/mid-ir-interpreter.test.ts
@@ -1,515 +1,0 @@
-import interpretMidIRCompilationUnit from '../mid-ir-interpreter';
-
-import {
-  ENCODED_FUNCTION_NAME_MALLOC,
-  ENCODED_FUNCTION_NAME_THROW,
-  ENCODED_FUNCTION_NAME_STRING_TO_INT,
-  ENCODED_FUNCTION_NAME_INT_TO_STRING,
-  ENCODED_FUNCTION_NAME_STRING_CONCAT,
-  ENCODED_FUNCTION_NAME_PRINTLN,
-  ENCODED_COMPILED_PROGRAM_MAIN,
-} from 'samlang-core-ast/common-names';
-import {
-  MIR_ZERO,
-  MIR_ONE,
-  MIR_EIGHT,
-  MIR_CONST,
-  MIR_NAME,
-  MIR_TEMP,
-  MIR_IMMUTABLE_MEM,
-  MIR_OP,
-  MIR_MOVE_TEMP,
-  MIR_MOVE_IMMUTABLE_MEM,
-  MIR_CALL_FUNCTION,
-  MIR_JUMP,
-  MIR_LABEL,
-  MIR_CJUMP_FALLTHROUGH,
-  MIR_RETURN,
-} from 'samlang-core-ast/mir-nodes';
-
-it('interpretMidIRCompilationUnit hello world test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [{ name: 'HW', content: 'Hello World!' }],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_NAME('HW')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('Hello World!\n');
-});
-
-it('interpretMidIRCompilationUnit string-int conversion test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_EIGHT], 'a'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_STRING_TO_INT), [MIR_TEMP('a')], 'b'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('b')], 'c'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('c')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('8\n');
-});
-
-it('interpretMidIRCompilationUnit failed string-int conversion test', () => {
-  expect(() =>
-    interpretMidIRCompilationUnit({
-      globalVariables: [{ name: 'HW', content: 'Hello World!' }],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_STRING_TO_INT), [MIR_NAME('HW')]),
-          ],
-        },
-      ],
-    })
-  ).toThrow('Bad string: Hello World!');
-});
-
-it('interpretMidIRCompilationUnit string concat test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [
-        { name: 'HW1', content: 'Hello World!' },
-        { name: 'HW2', content: ' Hi World!' },
-      ],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CALL_FUNCTION(
-              MIR_NAME(ENCODED_FUNCTION_NAME_STRING_CONCAT),
-              [MIR_NAME('HW1'), MIR_NAME('HW2')],
-              'a'
-            ),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('a')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('Hello World! Hi World!\n');
-});
-
-it('interpretMidIRCompilationUnit panic test', () => {
-  expect(() =>
-    interpretMidIRCompilationUnit({
-      globalVariables: [{ name: 'Ahh', content: 'Panic!' }],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_THROW), [MIR_NAME('Ahh')], 'a'),
-          ],
-        },
-      ],
-    })
-  ).toThrow('Panic!');
-});
-
-it('interpretMidIRCompilationUnit setup tuple and print test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_MALLOC), [MIR_CONST(16)], 'tuple'),
-            MIR_MOVE_IMMUTABLE_MEM(MIR_TEMP('tuple'), MIR_EIGHT),
-            MIR_MOVE_IMMUTABLE_MEM(MIR_OP('+', MIR_TEMP('tuple'), MIR_EIGHT), MIR_CONST(42)),
-            MIR_MOVE_TEMP(
-              'sum',
-              MIR_OP(
-                '+',
-                MIR_IMMUTABLE_MEM(MIR_TEMP('tuple')),
-                MIR_IMMUTABLE_MEM(MIR_OP('+', MIR_TEMP('tuple'), MIR_EIGHT))
-              )
-            ),
-            MIR_CALL_FUNCTION(
-              MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING),
-              [MIR_TEMP('sum')],
-              'sum-string'
-            ),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('sum-string')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('50\n');
-});
-
-it('interpretMidIRCompilationUnit jump test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_JUMP('foo'),
-            MIR_CALL_FUNCTION(
-              MIR_NAME(ENCODED_FUNCTION_NAME_THROW),
-              [MIR_OP('+', MIR_NAME('Ahh'), MIR_EIGHT)],
-              'a'
-            ),
-            MIR_LABEL('foo'),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('');
-});
-
-it('interpretMidIRCompilationUnit cjump test 1', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CJUMP_FALLTHROUGH(MIR_EIGHT, 'foo'),
-            MIR_CALL_FUNCTION(
-              MIR_NAME(ENCODED_FUNCTION_NAME_THROW),
-              [MIR_OP('+', MIR_NAME('Ahh'), MIR_EIGHT)],
-              'a'
-            ),
-            MIR_LABEL('foo'),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('');
-});
-
-it('interpretMidIRCompilationUnit cjump test 2', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CJUMP_FALLTHROUGH(MIR_ZERO, 'foo'),
-            MIR_RETURN(MIR_ZERO),
-            MIR_LABEL('foo'),
-            MIR_CALL_FUNCTION(
-              MIR_NAME(ENCODED_FUNCTION_NAME_THROW),
-              [MIR_OP('+', MIR_NAME('Ahh'), MIR_EIGHT)],
-              'a'
-            ),
-          ],
-        },
-      ],
-    })
-  ).toBe('');
-});
-
-it('interpretMidIRCompilationUnit dummy function call integration test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: 'dummy',
-          argumentNames: [],
-          mainBodyStatements: [MIR_RETURN(MIR_ZERO)],
-        },
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [MIR_CALL_FUNCTION(MIR_NAME('dummy'), []), MIR_RETURN(MIR_ZERO)],
-        },
-      ],
-    })
-  ).toBe('');
-});
-
-it('interpretMidIRCompilationUnit factorial function call integration test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: 'factorial',
-          argumentNames: ['n', 'acc'],
-          mainBodyStatements: [
-            MIR_CJUMP_FALLTHROUGH(MIR_OP('==', MIR_TEMP('n'), MIR_ZERO), 'l_RETURN_ACC'),
-            MIR_CALL_FUNCTION(
-              MIR_NAME('factorial'),
-              [MIR_OP('-', MIR_TEMP('n'), MIR_ONE), MIR_OP('*', MIR_TEMP('acc'), MIR_TEMP('n'))],
-              'dummy'
-            ),
-            MIR_RETURN(MIR_TEMP('dummy')),
-            MIR_LABEL('l_RETURN_ACC'),
-            MIR_RETURN(MIR_TEMP('acc')),
-          ],
-        },
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_CALL_FUNCTION(MIR_NAME('factorial'), [MIR_CONST(4), MIR_ONE], 'a'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('a')], 'b'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('b')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('24\n');
-});
-
-it('interpretMidIRCompilationUnit factorial function reference call integration test', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: 'factorial',
-          argumentNames: ['n', 'acc'],
-          mainBodyStatements: [
-            MIR_CJUMP_FALLTHROUGH(MIR_OP('==', MIR_TEMP('n'), MIR_ZERO), 'l_RETURN_ACC'),
-            MIR_CALL_FUNCTION(
-              MIR_NAME('factorial'),
-              [MIR_OP('-', MIR_TEMP('n'), MIR_ONE), MIR_OP('*', MIR_TEMP('acc'), MIR_TEMP('n'))],
-              'dummy'
-            ),
-            MIR_RETURN(MIR_TEMP('dummy')),
-            MIR_LABEL('l_RETURN_ACC'),
-            MIR_RETURN(MIR_TEMP('acc')),
-          ],
-        },
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('name', MIR_NAME('factorial')),
-            MIR_CALL_FUNCTION(MIR_TEMP('name'), [MIR_CONST(4), MIR_ONE], 'a'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('a')], 'b'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('b')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('24\n');
-});
-
-it('interpretMidIRCompilationUnit binary expression test 1', () => {
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('one', MIR_ONE),
-            MIR_MOVE_TEMP('v', MIR_OP('-', MIR_ONE, MIR_TEMP('one'))),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('0\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('/', MIR_ONE, MIR_ONE)),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('1\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('%', MIR_ONE, MIR_EIGHT)),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('1\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('^', MIR_ONE, MIR_ZERO)),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('1\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('<', MIR_ONE, MIR_ZERO)),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('0\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('<=', MIR_ONE, MIR_ZERO)),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('0\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('>', MIR_ONE, MIR_ZERO)),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('1\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('>=', MIR_ONE, MIR_ZERO)),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('1\n');
-
-  expect(
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [
-            MIR_MOVE_TEMP('v', MIR_OP('!=', MIR_ONE, MIR_TEMP('undefined_default_to_0'))),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_INT_TO_STRING), [MIR_TEMP('v')], 's'),
-            MIR_CALL_FUNCTION(MIR_NAME(ENCODED_FUNCTION_NAME_PRINTLN), [MIR_TEMP('s')]),
-            MIR_RETURN(MIR_ZERO),
-          ],
-        },
-      ],
-    })
-  ).toBe('1\n');
-
-  expect(() =>
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [MIR_MOVE_TEMP('v', MIR_OP('/', MIR_ONE, MIR_ZERO))],
-        },
-      ],
-    })
-  ).toThrow('Division by zero!');
-
-  expect(() =>
-    interpretMidIRCompilationUnit({
-      globalVariables: [],
-      functions: [
-        {
-          functionName: ENCODED_COMPILED_PROGRAM_MAIN,
-          argumentNames: [],
-          mainBodyStatements: [MIR_MOVE_TEMP('v', MIR_OP('%', MIR_ONE, MIR_ZERO))],
-        },
-      ],
-    })
-  ).toThrow('Mod by zero!');
-});
